# Conflicts:
#	src/main/webapp/index.jsp

+++ src/main/java/admin/AdminUserModifyReg.java
@@ -18,11 +18,15 @@ public class AdminUserModifyReg implements AdminService{
     public void execute(HttpServletRequest request, HttpServletResponse response) {
         try {
         String msg = "수정되었습니다";
+        AdminDAO dao = new AdminDAO();
+        String imgFile_check = ".*[.](jpeg|jpg|bmp|png|gif|pdf)";
 
 
-      //  String path = "C:\\Users\\82108\\TingleProject\\src\\main\\webapp\\img\\";
             String path = request.getRealPath("/img/");
         int size = 1024 * 1024 * 10;
+
+
+
 
         MultipartRequest mr = null;
 
@@ -33,37 +37,48 @@ public class AdminUserModifyReg implements AdminService{
                     "UTF-8",
                     new DefaultFileRenamePolicy());
 
+            boolean check = dao.fileExtension_Check(imgFile_check,mr.getFilesystemName("profile_image"));
 
-        UserDTO userData = new UserDTO();
+            if(check){
+                UserDTO userData = new UserDTO();
 
-        userData.setId(Integer.parseInt(mr.getParameter("id")));
-        userData.setLogin_id(mr.getParameter("login_id"));
-        userData.setLogin_pw(mr.getParameter("login_pw"));
-        userData.setName(mr.getParameter("name"));
-        userData.setNickname(mr.getParameter("nickname"));
-        userData.setPhone_number(mr.getParameter("phone_number"));
-        userData.setEmail_address(mr.getParameter("email.address"));
-        if(mr.getFilesystemName("profile_image")!= null){
-            userData.setProfile_image(mr.getFilesystemName("profile_image"));
-        } else {
-            userData.setProfile_image(mr.getParameter("filecheck"));
-        }
-        userData.setJoin_type(Integer.parseInt(mr.getParameter("join_type")));
-       userData.setMembership(Integer.parseInt(mr.getParameter("membership")));
-        userData.setPersonal_type(mr.getParameter("personal_type"));
+                userData.setId(Integer.parseInt(mr.getParameter("id")));
+                userData.setLogin_id(mr.getParameter("login_id"));
+                userData.setLogin_pw(mr.getParameter("login_pw"));
+                userData.setName(mr.getParameter("name"));
+                userData.setNickname(mr.getParameter("nickname"));
+                userData.setPhone_number(mr.getParameter("phone_number"));
+                userData.setEmail_address(mr.getParameter("email.address"));
+                if(mr.getFilesystemName("profile_image")!= null){
+                    userData.setProfile_image(mr.getFilesystemName("profile_image"));
+                } else {
+                    userData.setProfile_image(mr.getParameter("filecheck"));
+                }
+                userData.setJoin_type(Integer.parseInt(mr.getParameter("join_type")));
+                userData.setMembership(Integer.parseInt(mr.getParameter("membership")));
+                userData.setPersonal_type(mr.getParameter("personal_type"));
 
-            if(mr.getFilesystemName("profile_image")!= null && mr.getFilesystemName("profile_image") != mr.getParameter("filecheck")){
-                Path overlapfile = Paths.get(path + mr.getParameter("filecheck"));
-                Files.delete(overlapfile);
+                if(mr.getFilesystemName("profile_image")!= null
+                        && mr.getFilesystemName("profile_image") != mr.getParameter("filecheck")
+                        && !mr.getParameter("filecheck").equals("exfire_member.png")){
+                    Path overlapfile = Paths.get(path + mr.getParameter("filecheck"));
+                    Files.delete(overlapfile);
+                }
+
+                dao.userModify(userData);
+
+                int id = userData.getId();
+
+                request.setAttribute("msg", msg);
+                request.setAttribute("adminUrl", "alert.jsp");
+                request.setAttribute("goUrl", "AdminUserDetail?id="+id);
+            } else {
+                msg = "파일형식이 잘못되었습니다";
+                request.setAttribute("msg", msg);
+                request.setAttribute("adminUrl", "alert.jsp");
+                request.setAttribute("goUrl", "AdminUserDetail?id="+ mr.getParameter("id"));
             }
 
-        new AdminDAO().userModify(userData);
-
-        int id = userData.getId();
-
-        request.setAttribute("msg", msg);
-        request.setAttribute("adminUrl", "alert.jsp");
-        request.setAttribute("goUrl", "AdminUserDetail?id="+id);
 
         } catch (IOException e) {
             throw new RuntimeException(e);
